+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This release contains several breaking changes which were necessary for the heal
 * Removed GeographicCoordinateException.Messages.COORDINATE_TYPE_NULL, as it can no longer be thrown
 * Changed GeographicCoordinateException.Messages from public to protected, as client code should not be using them.  The constants contained therein exist solely to keep the internal code and JUnit tests in sync.  The javadoc has been updated to indicate this.
 * Changed the Latitude/Longitude toString methods to return a degree/minute/second format, such as 12°34'56.789"N.  Likewise, Point's toString method now uses this format.
+* When using the Latitude(double) constructor to create a latitude of 0.0 (the Equator), the direction will be considered north.  Previously, it was considered south.  No client code should be impacted because it's never correct to examine the direction when talking about the Equator, since it is neither north nor south.
+* When using the Longitude(double) constructor to create a longitude of 0.0 (the Prime Meridian), the direction will be considered east.  Previously, it was considered west.  No client code should be impacted because it's never correct to examine the direction when talking about the Prime Meridian, since it is neither east nor west.
 * Integrated with the latest version of the BuildScripts project (https://github.com/kloverde/BuildScripts), adding findbugs integration, JUnit integration and new release packaging:  one archive contains all zips and the readme, changelog and license.
 * Various javadoc updates